feat: rewrite local storage service to functional component
