+++ src/__mocks__/index.ts
@@ -1,0 +1,3 @@
+import { reqestMockData } from './responseMocks';
+
+export { reqestMockData };
